Panda Picks/requirements.txt
numpy==1.21.2
pandas>=1.3.3
requests==2.26.0
flask
streamlit
altair
tabula-py